Add DPDK version support.
DPDK APIs paths and function are changing time to time for each release.
The test scripts are based only on OVS release. It is impossible
to specify what version of DPDK is using while running the scripts
Now adding the support for specifying DPDK version.
DPDK version can be specified in the environment file
'DPDK_VER'. It can be 16.04, 16.07, ... By default the version
is 16.11

Signed-off-by: Sugesh Chandran <sugesh.chandran@intel.com>